Refactors the routes into separate files and adds routes for sentence/paragraph length counts
